+++ packages/atproto/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.7.2
 
 - Supported `com.atproto.labels.queryLabels` and `LabelsService`. ([#265](https://github.com/myConsciousness/atproto.dart/issues/265))
+- Supported `com.atproto.labels.subscribeLabels` on `LabelsService`. ([#340](https://github.com/myConsciousness/atproto.dart/issues/340))
 
 ## v0.7.1